lambda/requirements.txt
boto3==1.34.0
Pillow==11.3.0
